Remove ConfigTime file and combine it to ConfigSchedule
This way, if someone wants to make a change/add a new schedule, they can just go to one file instead of worrying about two.
