+++ common/src/index.ts
@@ -2,13 +2,10 @@ export * from './enums/SchoolType'
 export * from './enums/SpecialPeriod';
 
 export * from './config/InClassSchedule';
-export * from './config/InClassTimes';
 export * from './config/RemoteSchedule';
-export * from './config/RemoteTimes';
 
 export * from './interfaces/InputClass';
 export * from './interfaces/Period';
 export * from './interfaces/Schedule';
 export * from './interfaces/ScheduleInput';
-export *  from './interfaces/ConfigurationSchedule';
-export * from './interfaces/ConfigurationTimes';
+export * from './interfaces/ConfigurationSchedule';
