--- common/src/interfaces/ConfigurationTimes.ts
@@ -1,13 +1,0 @@
-import {SchoolType} from "..";
-import {SpecialPeriod} from '../enums/..';
-
-interface Times {
-    period: number | SpecialPeriod,
-    from: string,
-    to: string
-}
-
-export interface ConfigurationTimes { 
-    [SchoolType.HIGH_SCHOOL]: Times[],
-    [SchoolType.MIDDLE_SCHOOL]: Times[]
-};
